Add Turbine refurb
Adds investment option for turbine. It consumes heat generated by P2 and converts to electricity at an assumed efficiency of 0.4. The turbine cannot generate electricity without heat from P2 and thus investments will not occur unless P2 is also invested in.

NB: PEF_loc needs to be adjusted to account for the fact that the turbine generates electricity which decreases PEF.