Modify PV equation
Input files were in kWh/kW_capacity so this commit adjusts the equations to reflect that. Constraints are now formulated so that installed capacity is less than roof area times the PV capacity density in kW per m2. A similiar change is made to eq_PV which simplifies the expression.