Big Code Clean Up
Renamed many variables to more consistent naming, modified almost every
line of code in the FED_MAT_Main.m script. Cleaned up the code in both
GAMS and Matlab so that it was more concise and easier to understand
including removal of temp variables, legacy code, output gdx files that
weren't used etc. One side affect was that a bug relating to the use of
the command "ord(m)" was discovered that led to a need to comment out
code having to do with AH actual costs. This needs to be reformulated if
we want it correct. Code runs as is now.