Attempt at fixing BTES SoC
Ensures that only values for a single hour are used in BTES. Should be the first hour in the current simulation. AKA h-1 in the coming simulation for initial stored energy.

If BTES_Den and BTES_Sen for some reason haven't been set at a 'level' in GAMS a try - catch combination in read_GtoM creates a new struct where all storages are empty and returns instead.