Changed BID for AH, fixed warnings and hard coded files
The previous commit did not run on a new computer so in the process of getting it running a lot of warnings were found in FED_MAT_MAIN and get_synthetic_baseline_load_data. I have fixed so that the code runs now with the current version of Input_Dispatch_model folder on Dropbox. These changes need to be checked by David.